修复了“每日签到数据统计”中的bug

+++ PhdChorus/ZivDBManager+Statics.h
@@ -11,11 +11,11 @@
 @interface ZivDBManager (Statics)
 
 // part = S | A | T | B
-- (void)part:(nonnull NSString *)part memberCountOfHighPart:(nonnull int *)highPartCount andLowPart:(nonnull int *)lowPartCount;
-- (void)part:(nonnull NSString *)part memberCountOfZhongguancun:(nonnull int *)zhongguancunCount andYanqi:(nonnull int *)yanqiCount;
-- (void)part:(nonnull NSString *)part absenceCountOfHighPart:(nonnull int *)highPartCount andLowPart:(nonnull int *)lowPartCount inAttendanceTable:(nonnull NSString *)tableName;
-- (void)part:(nonnull NSString *)part attendanceCountOfHighPart:(nonnull int *)highPartCount andLowPart:(nonnull int *)lowPartCount inAttendanceTable:(nonnull NSString *)tableName;
-- (void)part:(nonnull NSString *)part attendanceCountOfZhongguancun:(nonnull int *)zhongguancunCount andYanqi:(nonnull int *)yanqiCount inAttendanceTable:(nonnull NSString *)tableName;
+- (void)part:(nonnull NSString *)part memberCountOfHighPart:(nonnull NSInteger *)highPartCount andLowPart:(nonnull NSInteger *)lowPartCount;
+- (void)part:(nonnull NSString *)part memberCountOfZhongguancun:(nonnull NSInteger *)zhongguancunCount andYanqi:(nonnull NSInteger *)yanqiCount;
+- (void)part:(nonnull NSString *)part absenceCountOfHighPart:(nonnull NSInteger *)highPartCount andLowPart:(nonnull NSInteger *)lowPartCount inAttendanceTable:(nonnull NSString *)tableName;
+- (void)part:(nonnull NSString *)part attendanceCountOfHighPart:(nonnull NSInteger *)highPartCount andLowPart:(nonnull NSInteger *)lowPartCount inAttendanceTable:(nonnull NSString *)tableName;
+- (void)part:(nonnull NSString *)part attendanceCountOfZhongguancun:(nonnull NSInteger *)zhongguancunCount andYanqi:(nonnull NSInteger *)yanqiCount inAttendanceTable:(nonnull NSString *)tableName;
 
 - (nullable NSString *)exportAttendanceTableOfPart:(nonnull NSString *)part from:(nonnull NSString *)startTime to:(nonnull NSString *)endTime;
